CustomFieldsDraft to interface. #942
