+++ sphere-models/src/it/java/io/sphere/sdk/carts/commands/CartCreateCommandTest.java
@@ -5,10 +5,7 @@ import io.sphere.sdk.carts.*;
 import io.sphere.sdk.models.Address;
 import io.sphere.sdk.test.IntegrationTest;
 import io.sphere.sdk.test.JsonNodeReferenceResolver;
-import io.sphere.sdk.types.CustomFieldsDraft;
-import io.sphere.sdk.types.FieldDefinition;
-import io.sphere.sdk.types.StringFieldType;
-import io.sphere.sdk.types.TypeFixtures;
+import io.sphere.sdk.types.*;
 import io.sphere.sdk.types.commands.TypeDeleteCommand;
 import io.sphere.sdk.types.queries.TypeQuery;
 import org.junit.BeforeClass;
