+++ sphere-models/src/main/java/io/sphere/sdk/types/CustomFieldsDraft.java
@@ -1,86 +1,43 @@
 package io.sphere.sdk.types;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import io.sphere.sdk.json.SphereJsonUtils;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.models.ResourceIdentifier;
 
 import javax.annotation.Nullable;
 import java.util.Map;
-import java.util.stream.Collectors;
 
-import static io.sphere.sdk.utils.SphereInternalUtils.immutableCopyOf;
+import static io.sphere.sdk.types.CustomFieldsDraftImpl.mapObjectToJsonMap;
 import static java.util.Objects.requireNonNull;
 
-/**
- * @see Custom
- * @see CustomFieldsDraftBuilder
- */
-public class CustomFieldsDraft {
-    private final ResourceIdentifier<Type> type;
+@JsonDeserialize(as = CustomFieldsDraftImpl.class)
+public interface CustomFieldsDraft {
+    ResourceIdentifier<Type> getType();
+
     @Nullable
-    private final Map<String, JsonNode> fields;
+    Map<String, JsonNode> getFields();
 
-    CustomFieldsDraft(@Nullable final String typeId, @Nullable final String typeKey, final Map<String, JsonNode> fields) {
-        this.type = ResourceIdentifier.ofIdOrKey(typeId, typeKey);
-        this.fields = immutableCopyOf(fields);
+    static CustomFieldsDraft ofCustomFields(final CustomFields custom) {
+        return ofTypeIdAndJson(custom.getType().getTypeId(), custom.getFieldsJsonMap());
     }
 
-    public static CustomFieldsDraft ofCustomFields(final CustomFields custom) {
-        return CustomFieldsDraft.ofTypeIdAndJson(custom.getType().getTypeId(), custom.getFieldsJsonMap());
-    }
-
-    public static CustomFieldsDraft ofTypeIdAndJson(final String typeId, final Map<String, JsonNode> fields) {
+    static CustomFieldsDraft ofTypeIdAndJson(final String typeId, final Map<String, JsonNode> fields) {
         requireNonNull(typeId);
-        return new CustomFieldsDraft(typeId, null, fields);
+        return new CustomFieldsDraftImpl(typeId, null, fields);
     }
 
-    public static CustomFieldsDraft ofTypeIdAndObjects(final String typeId, final Map<String, Object> fields) {
+    static CustomFieldsDraft ofTypeIdAndObjects(final String typeId, final Map<String, Object> fields) {
         final Map<String, JsonNode> fieldsJson = mapObjectToJsonMap(fields);
         return ofTypeIdAndJson(typeId, fieldsJson);
     }
 
-    public static CustomFieldsDraft ofTypeKeyAndJson(final String typeKey, final Map<String, JsonNode> fields) {
+    static CustomFieldsDraft ofTypeKeyAndJson(final String typeKey, final Map<String, JsonNode> fields) {
         requireNonNull(typeKey);
-        return new CustomFieldsDraft(null, typeKey, fields);
+        return new CustomFieldsDraftImpl(null, typeKey, fields);
     }
 
-    public static CustomFieldsDraft ofTypeKeyAndObjects(final String typeKey, final Map<String, Object> fields) {
+    static CustomFieldsDraft ofTypeKeyAndObjects(final String typeKey, final Map<String, Object> fields) {
         final Map<String, JsonNode> fieldsJson = mapObjectToJsonMap(fields);
         return ofTypeKeyAndJson(typeKey, fieldsJson);
-    }
-
-    /**
-     * use {@link #getType()} instead
-     * @return id of the type to set
-     */
-    @Deprecated
-    @Nullable
-    public String getTypeId() {
-        return type.getTypeId();
-    }
-
-    /**
-     * @deprecated use {@link #getType()} instead
-     * @return key of the type to set
-     */
-    @Deprecated
-    @Nullable
-    public String getTypeKey() {
-        return type.getKey();
-    }
-
-    public ResourceIdentifier<Type> getType() {
-        return type;
-    }
-
-    @Nullable
-    public Map<String, JsonNode> getFields() {
-        return fields;
-    }
-
-    private static Map<String, JsonNode> mapObjectToJsonMap(final Map<String, Object> fields) {
-        return fields.entrySet().stream()
-                .collect(Collectors.toMap(entry -> entry.getKey(),
-                        entry -> SphereJsonUtils.toJsonNode(entry.getValue())));
     }
 }
